[Model] Extract GatedDeltaNetAttention into shared layer for Qwen3Next and Qwen3.5
